Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-RESTX==1.3.2
Flask-CORS==6.0.1
pandas
scikit-learn